Stabilize widget auto tests which use evaluateJavaScriptSync()
Interrupt the waiting on the event loop in CallbackSpy::waitForResult()
to not block calls on the UI thread.

It is necessary because RenderProcessHost creates channel for mojo
communication on the UI thread. QWebEnginePage::runJavaScript() needs
this channel because the JavaScript is executed in the render process.

If evaluateJavaScriptSync() is called before the mentioned channel is created,
the JavaScriptExecuteRequest mojo message might not be sent because the
wait would block the thread.

Change-Id: Ic5bb5a6fde02717cec49dcf9e458f2eaac09eacf
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>