PdfLinkModel: support the remaining link action types
In practice, some PDFs contain links for which FPDFAction_GetType()
reports PDFACTION_UNSUPPORTED and yet we can successfully get the same
information as if it were PDFACTION_GOTO.  For example
https://www.openexr.com/documentation/TechnicalIntroduction.pdf
contains some of these.

PDFACTION_URI is another web link representation: for example in
https://www.w3.org/WAI/WCAG21/working-examples/pdf-links/links.pdf

PDFACTION_LAUNCH and PDFACTION_REMOTEGOTO will be trickier to support in
actual viewer applications, but at least we should provide the file path
to open, as a URL.

Log similar warnings each time an invalid aspect of a link is
encountered, into the qt.pdf.links logging category.

Change-Id: I7ca24baa10524611e0705e89a07906ab7b74dadb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>