Add tst_MultiPageView::navigation() test
The navigation() test is a high-level one to verify that the navigation
stack works and that the TableView jumps to the expected locations
inside the document when we go to pages, click links and use the
forward() and back() methods. It's similar to the internalLink()
test, but more programmable.

Change-Id: I337e2ae283213238288a172d1bab3df07a1395e0
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 2d8c8ac02e8afed8d99996e3d53cacd982badfeb)
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>