PdfMultiPageView: add workaround to deal with scale change
TableViewExtra positionViewAtCell() / positionViewAtRow() merely set
contentY (also contentX if needed); but if the delegates are changing
in size, it goes undetected until one tries to scroll into a certain
range where Flickable detects that originY needs to be adjusted.
Getting TableView to do scheduleRebuildTable(RebuildOption::All)
forces it to start over with delegate sizes too, as if the model was
a QAIM that emitted modelReset().  Since our model is just a number,
one way we can trigger that same behavior is to reassign it.

This won't be needed when we have the newer TableView with
positionViewAtCell() as a built-in function.

Change-Id: I18e305b4ab19302464e289e2fc403b97346b1faa
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>