QPdfDocument::getSelection(): fix off-by-one character selection
When mouse-dragging to select text, the selection always contained one
more character than the geometry (set of polygons) was showing.
Conversely, it was not actually possible to select a single character:
you'd always get at least two, even though only one looked like it was
selected.

Applied a similar fix as the one in hitTest() (in
8d13facbd88d821cc89b21f43708cc1a81ac79f3): check the bounding box of the
last character, and if the given end position (where the mouse is being
dragged to) is closer to the right edge than to the left, increment the
ending character index, to point to the character after (which is
probably an invisible newline or other terminator).  This fix might not
be so great if the text doesn't run left-to-right, but it works for
Western text in normal orientation at least.

Fixes: QTBUG-83685
Change-Id: I099424cd334fa38986445c134c82580a0307c2fe
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>