Fix upside-down render passes with software scenegraph (again)
Originally, in 2017, this same issue was fixed by vertically mirroring the
texture used for the render pass (a311b7bc9d). At this time we were using
QSGSoftwareInternalImageNode for the scenegraph node and QSGSoftwareLayer for
the layer texture.

Then, in 2018, the implementation of QSGSoftwareInternalImageNode was changed to
automatically mirror its texture IF the texture is a QSGSoftwareLayer. This was
not a problem because the layer was also changed to mirror its contents, i.e.
the meaning QSGSoftwareLayer::setMirrorVertical was inverted. See 10c56615d in
qtdeclarative.

Finally, in 2019, we switched to using QSGSoftwareImageNode for the scenegraph
node of a render pass. This caused the bug to return, since this node does not
automatically mirror QSGSoftwareLayer textures. Fix it again by setting
setMirrorVertical(false) on the layer.

The default scenegraph is not affected by this change because the default image
node uses the layer's normalizedTextureSubRect to automatically undo the effect
of mirroring, i.e. mirroring the default layer has no effect on output.

Fixes: QTBUG-83785
Change-Id: I3594c0c7f5364799af1f543d1e4e3de52a25cd2b
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>