Fix path validation in FilePickerController::accepted()
In some cases QUrl::toLocalFile() does wrong path conversion on
Windows. Therefore we have to convert the file URL to absolute path
in another way and use FilePath::IsAbsolute() for validating that.

Also stabilize WebEngineViewSingleFileUpload tests. They didn't wait
for html title change when the expected value was the same as the
default title.

Change-Id: Ica7798a299f9b28657afeeccccfba3fdecc515e2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>