Fix scrollbar when releasing mouse button outside of window on Windows
When releasing mouse button outside of window on Windows, an extra
QEvent::Leave is sent before the QEvent::MouseButtonRelease. This
extra Leave event resets the pointer to the last interacted scrollbar
(last_scrollbar_under_mouse_) in Chromium. Leave events should not be
sent until the last mouse button is released.

Fixes: QTBUG-70437
Pick-to: 5.15
Change-Id: Ibf28b3ca0fd82feb5abf91ec851d92db0a8657f0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>