Fix crash on WebContentsDelegateQt::CloseContents()
CloseContents() triggers a windowCloseRequested() signal. The handler of
the signal might close the QWebEngineView and it would result the
destruction of the WebContentsDelegateQt. Thus, any operation on
WebContentsDelegateQt is not safe after a
WebContentsAdapterClient::close() call.

For example, calling QWebEngineView::close() on
QWebEnginePage::windowCloseRequested() would result crash because it
sets life cycle state to discarded which destructs web contents.

Fixes: QTBUG-84355
Change-Id: I5da404bc9b1923cc19085ee899b550da49d1416b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>