Add some info when project is misconfigured
We support calling make on not configured properly qtwebengine and qtpdf.
This avoids issues when there is no make file for top level
build and user does not understand why. To do that we 'share'
error messages between configure and make calls. However to avoid
duplicating tests and error messages 'configure test conditions' have to
match 'error messages' in support.pri.

In case we miss some error message just ask for reporting the bug,
instead of still trying to build the thing.

Change-Id: I23320afa5174b5f33871c8cd8739ceea8e3a169d
Reviewed-by: Michael Brüning <michael.bruning@qt.io>