Fix incorrect doc for acceptNavigationRequest and add test
The function is called for all navigation requests. The original bug
report was most likely due to the use relative links on a page without
a base url, which indeed results in no call to acceptNavigationRequest
being made. But this has nothing to do with local vs remote URLs, it's
rather that relative links, which are relative to nothing, are not
valid links and there thus cannot be any navigation request either.

Task-number: QTBUG-48435
Change-Id: I08bd0c86d67bf1dd1d7662468321777254a7db0b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>