Fix swap condition in DisplayGLOutputSurface::updatePaintNode
The buffer pointers m_middleBuffer and m_frontBuffer are swapped in
updatePaintNode on the condition that m_middleBuffer is non-null and
it's serviceId non-zero. But neither m_middleBuffer nor it's serviceId
is immediately cleared, leaving the possibility that a subsequent call
to updatePaintNode could swap the buffer pointers again if this call
occurs before m_middleBuffer is cleared in swapBuffersOnVizThread. The
m_taskRunner pointer is however cleared immediately and therefore the
subsequent call to updatePaintNode will trigger a segmentation fault.

Since m_taskRunner precisely tracks the condition when we should or
should not swap, change the swap condition to simply require that
m_taskRunner is non-null.

Fixes: QTBUG-85817
Change-Id: Iad89bb0f4de7c0c151d5c5bd63ac74ab6cdd9087
Reviewed-by: Michael Brüning <michael.bruning@qt.io>