Introduce qml script collection and make webscript as basic type
In widget apis webenginescripts are value types and are managed
by webenginescriptcollection. Introduce same concept in qml and
make qquickwebenginescript a value type and reuse core userscript
as private implementation. Rewrite webenginscript list handling.

This patch is half baked since it does not include docs and tests
for new api, however aim is to move script classes to core and reuse
those in qml land. Therefore, new class introduced here is going to
be removed in follow up patches, so all the missing parts will be added
later. A new way of managing scripts in qml is as follows:

  * using collection with javascript dictionaries

    var script = { name: "FOO"
                   sourceUrl: Qt.resolvedUrl("foo.js"),
                   injectionPoint: WebEngineScript.DocumentReady }

    webEngineView.userScripts.collection = [ script1, script2 ];

  * using collection with webscript basic type

    var script = WebEngine.script()
    script.name = "FOO"
    webEngineView.userScripts.collection = [ script ];

  * using fine grain user script collection api with basic type

    var script = WebEngine.script()
    script.name = "FOO"
    webEngineView.userScripts.insert(script)

Of course new api can be extended and we can provide more convince
overloads.

Note the main motivation here is to enable reuse webenginescript object
created in c++ land, which is now passed as value in follow up
patches.

This changes reuses private apis of qml and will most likely require
further changes when QTBUG-82443 is completed.

[ChangeLog] WebEngineScript is a basic value type in qml, it is no
longer declarative way of creating it, instead use WebEngine.script()

Change-Id: I6a0ac3607e4522ccaefcec0a7d2986577d7e7024
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>