Fix assert in WebContentsAdapter::devToolsFrontendDestroyed
Can be reproduced with quicknanobrowser by

  1. Open devtools
  2. Switch to OTR mode

During the switch to OTR mode, the inspected WebContentsAdapter is
destroyed and recreated. DevToolsFrontendQt is supposed to call
devToolsFrontendDestroyed on the old WebContentsAdapter, but it
accesses this adapter through QQuickWebEngineViewPrivate's
webContentsAdapter() which has already been changed to the new value,
so devToolsFrontendDestroyed is actually called on the wrong object.

Fix by remembering the WebContentsAdapter pointer in the
DevToolsFrontendQt constructor.

Change-Id: I17004ebb1a392399ef3e189c433d3f4b3f05bfd3
Reviewed-by: Tamas Zakor <ztamas@inf.u-szeged.hu>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>