Enable webrtc logging and the corresponding WebUI
The corresponding WebUI is chrome://webrtc-logs

It only makes sense with the hangout services extension. It seems to be
only useable with meet.google.com. The behavior is same for Chrome.

Uploading logs to Google is disabled in the Chromium patch. It is an
extension API functionality. The WebUI only lists the previously
generated logs.

Pulls in the following changes:
1dda5314b02 Enable webrtcLoggingPrivate extension API for hangout extension
7d71aca116d FIXUP: Enable webrtcLoggingPrivate extension API for hangout extension

Change-Id: I455fce7c8081e71967e55ab1f889df6ef91ed253
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>