Transform/Rotate qml popups in the case of multiwindow support
This is ugly attempt to cover the case for different
screen orientations with qml popups on system with
windowing system in form of:

Item {
   anchors.fill: parent
   rotation: 90
   WebEngineView {
      anchors.fill: parent
   }
}

This implemented by adding special transformation for rotation
for quickwindow delegate.

Do not support move or resize in that case since chromium is not aware
of this transformation anyway.

Fixes: QTBUG-63888
Fixes: QTBUG-75893
Task-number: QTBUG-73244
Task-number: QTBUG-89358
Pick-to: 5.15
Change-Id: I0a116e0b85dc17e6ad517829a2db1d3f192d32ce
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>