Implement PluginServiceFilterQt
Let Chromium know about the status of plugin availability from
WebEngineSettings. This way it can decide whether the response is
a download and it has more benefits:
- It doesn't have to start a new load request to download a PDF file,
it just treats the original response accordingly.
- Fixes websites which are protected from cross-domain requests (e.g. by
checking the Referer header) and/or redirecting requests for PDFs.
Calling DownloadManager "manually" and not passing the original request
headers did not work when the server relied on them.

Task-number: QTBUG-78114
Change-Id: I8cfa90c211418001c60c4b2f0f8818ee453101fc
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>