Abort findText also right on explicit navigation request
Current implementation cancels current findText only when new navigation
is accepted (since it may be rejected all together), so it's not guaranteed
that user code will not receive completion callback if it arrives after
explicit load/setContent/setHtml but before acceptNavigationRequest.
For explicit navigation it doesn't make sense to wait until it's
accepted, since it's only exposed there just for consistency, and an
expectation for findText is that it should be canceled on new navigation.

Fixes: QTBUG-61887
Change-Id: Ia2e19df3b5712e6b5426443d1bce6b205e186668
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>