Remove tracking of frame which load error page
It was added to suppress progress notification for error page load
after failure, but since error page load is reported as a new navigation
(which clears list of tracked frames), it was actually doing the
opposite thing. The only situation where it suppresses progress is when
navigation was not finished (due to invalid domain or network error),
but in this case it was real progress change for whole load which should
propagate further.

Change-Id: Ifd1d681fb5c6495fb3afdc4247364afb4472c959
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>