Depend on QCoreApplication::startingUp() for checking existence of app
When using QtWebView with WebEngine then it will hook the call to
QtWebEngine::initialize() to the start up of the application object which
means it will output the warning because it already exists. However
there is still time at this point to set what is needed because it is
still being initialized. So by checking if the application is running
(i.e. !startingUp()) then we can be safe in knowing that it is still
able to do the initialization.

Change-Id: I8c5d8808b4b09e1e7bbf4be52e5efc0786ce1472
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>