Fix not working ninja multi config on coin
On coin we call cmake --build . and we do not specify
the config expecting that all configuration are build.
However, when using the custom command for ninja build
dependencies are incorrectly resolved by cmake and
the ninja step is only executed in "default" config context.
This breaks debug builds. Cmake 3.20 fixes the issue
by adding genex to OUTPUT and support for OUTPUT_CONFIG.

Create workaround for now and revert it when minimum version
of cmake is bumped to 3.20

Change-Id: Iafaedb5f605ed9195dd4ecaa34d2dff19b7225a6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>