Fix QWebEngineDownloadRequest signals
downloadFileNameChanged and downloadDirectoryChanged signals were removed in
    08375ee7 Rename QWebEngineDownloadItem -> QWebEngineDownloadRequest.
Reintroduce these signals because they are still part of the API.
Also test these in Widgets auto test.

QQuickWebEngineDownloadItem::pathChanged were also removed in
    75393077 Remove obsolete, deprecated api from qwebenginedownloaditem.
Remove the pathChanged signal checks from the test because it was deprecated.

The test tst_download.qml does not have quickcontrols1 dependency.
Update it in the project file and re-enable the test.

Change-Id: I702e04dfe2179beafc558a628bfa07f4090fede2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>