Replace testsupport's QQuickWebEngineErrorPage with isErrorPage
Merge and unify handling of loading started/finished for quick
and widgets by removing separate quick's type for monitoring
error page load and replace it with an isErrorPage method
in WebEngineLoadRequest to indicate, that load ended with an error
page being displayed. Effectively this only slightly changes when
loading finished gets emitted: now the signal is just postponed until
error page is finished loading after initial failure.

[ChangeLog][QWebEngineQuick][WebEngineLoadingInfo] New property
'isErrorPage' which indicates that the load resulted in an error page

Change-Id: I3e59dc488429d776f7c8e083b6d0489fb30a65fc
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>