Port qqc2 delegates to Qt6 and cmake build
It seems that dialogs in qqc2 are still not native windows,
but embedded qquickitems. We used qqc1 dialogs so they look good
on desktop and behave like dialogs. With Qt6 qcc1 are gone,
therefore switch dialogs to qqc2. Unfortunately they no longer
behaves as dialogs (controls are just placed in the middle of a
parent window), but still it is better than nothing.

Pick-to: 6.2
Task-number: QTBUG-93666
Change-Id: I282ac9c8228eb0bfeefce9b8ef4f657f02a597e7
Reviewed-by: Balazs Egedi <egedib@inf.u-szeged.hu>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>