Support HTML5 <datalist> element
The datalist uses Chromium's autofill component to fetch and filter
predefined options in the list and autocomplete the field with the
selected option. Autofill component is added to build and bound to
WebEngine. All the unnecessary autofill features for datalist are
supposed to be disabled: payment/credit card support, password manager,
save profile data, store suggestions in database etc.

Custom popups for the dropdown are implemented for Widget and Quick.
Scrolling in dropdown is not implemented in this change.

Fixes: QTBUG-54433
Change-Id: I155d02d6dbc9d88fbca4bc5f55b76c19d0ba7a9d
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 19c6b0a22f20a33c8c1fd4bdea29468fa8b4e10e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>