Fix assertion in getCurrentKeyboardLayout()
FATAL:xkb_keyboard_layout_engine.cc(640)] Keymap file failed to load: hu-us

If there are multiple layouts provided (eg. "hu,us") and there is no
layout variant, we appended the second layout as a variant (eg. "hu-us").
This happens because strtok() caches the pointer of the next token from
its previous call and uses that when processing a null string.

Pick-to: 6.2
Change-Id: Ic1d8bd14f95bda65a9fba82c1df45af306c4445b
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>