Remove check for profile directory from favicon service factory
Since the directory itself in initialized asynchronously this check
may lead to desynchronization with how profile is considered to be off
the record and what it really is. Subsequently, with a possible wrong
initialization of a favicon service with invalid history service for
non-otr profile.

Fixes: QTBUG-95580
Change-Id: I66fbb7914a303650490e1577989beee5cd883037
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 498c4ac362cb6cbee8a30c8cc46fa4253bf93ee0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>