Disable examples compilation on qemu
Cmake builds when cross compiling have 3 different sets
of qt libs:

 * sysroot
 * host
 * staging ones (target)

However unlike qmake cmake does not support rpath-link therefore libs
during the linking can be resolved to the sysroot instead of
staging (target) ones, meaning we compile against newer headers
and link with older sysroot's lib.

Luckily this only happens when cmake is not able to set full absolute
path to the library in staging dir, which in turn depends on rpath of
the library. If rpath start with ORIGIN cmake will make absolute
path and all implicit dependency libs are pulled from that lib location.

However, during examples builds libs for given module are not installed
yet, meaning there is no ORIGIN in rpath libraries so no absolute path
meaning that all implicit dependencies in result are pulled from
sysroot. This leads to linker errors in best scenario or to faulty
linked application.

The workaround introduced in QTBUG-86533 is to add all implicit
dependencies as explicit, but in case of webengine it means adding log
list of implicit dependencies to every single example.

Simply disabled build of examples as workaround.

Note tests are not currently affected by that issue on coin, since they
are built after installing module libs, which is not the case for
examples.

Note this issue is affecting us now, since external project is gone and
we do not install libs implicitly during module build stage.

Task-number: QTBUG-86533
Change-Id: Ica29cffcfbfcf66912af60ba4f8e88eca3af6f78
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>