Fix crash on WebEngine(View|Profile).userScripts.collection get
To create a javascript list on 'collection' method qml engine is needed
for a qml scope, but it's only available for objects, which are created
by the QML Engine itself. The only object in hierarchy of classes is an
actual WebEngineView, so get it from there on collection init.
Also implement delayed initialization of scripts collection for the
usage of default profile through WebEngine singleton, and for the reason
that it's not really needed until really asked by user code.

Pick-to: 6.2 6.3
Fixes: QTBUG-96597
Change-Id: I61e76652a5d0fd5609070fd541816503908f2dc8
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>