Fix WebChannel when JavaScript is disabled
WebChannel was not working in ApplicationWorld with JavaScript disabled
in MainWorld, because WebChannelIPCTransport::DidClearWindowObject() is
called only when the window object was cleared in the main world. Moving
the WebChannelTransport installation logic to DidCreateScriptContext()
works in other worlds, so fixes the problem.

Task-number: QTBUG-88875
Change-Id: Ia75613b66a1e049f617f0664684b153b6875e9de
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 70abc0b2eebc777309d372203df5ca1e785402d8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>