Doc: Fix QDoc warnings for licensing documentation
licenses.py takes the value of the URL attribute directly from
Chromium's license attribution files, and uses it as a link
target for QDoc's \l command.

Some of the URLs are not valid, however. Add \externalpage entries
for the invalid ones to either direct to the correct URL or to
expand to an empty '#' link.

Change-Id: I06e9c54bceb5465c89de115cb8cd8aa0e75e9c51
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 2e268202d4f1ae14698aff3eabb1d60e6d04128d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>