Fix pinch gesture
Pinch gesture on a touchpad is expected to zoom-in and zoom-out. It has
been broken since the pinch gestures are routed because for routing the
event target has to be found. The event target is only tried to be found
on a pinch begin gesture.

As a fix, handle Qt::BeginNativeGesture and Qt::EndNativeGesture events
too.

Fixes: QTBUG-96930
Pick-to: 6.2 5.15
Change-Id: Ic8fe5bee933b5e0fbc8f5ba6234363a0a625648d
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>