Use new x11 native interface to query xdisplay
The call to nativeResourceForScreen can return egl native display
handle if called with "display", since it gets mapped to XLibDisplay
but egl device intergration can return EGLDisplay.
This is not the case for nativeResourceForIntegration.

Use new native QX11Application interface to query for display,
note there is not need to use nativeResourceForScreen for "display"
anyway since qt does not support multiple x connections to different
displays.

This fixes places where code queries "display" and bails out
if it is null.

Fixes: QTBUG-97472
Pick-to: 6.2 6.2.2
Change-Id: Ibc5f8f96e612389bfa24a81268202c9e47a7580b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>