Do not cancel custom tooltips by QWebEngineView custom tooltip handler
Tooltips should disappear when chromium requests it and it is done by an
explicit call of QToolTip::hideText() (see QTBUG-64933).

This, however would also hide a custom "global" tooltip because the
custom tooltip event notifies QWebEngineView to remove its own tooltip
to avoid more than one in a window. The QWebEngineView's tooltip handler
was override and it resulted custom tooltips to hide themselves.

As a fix, remove the custom tooltip handler from QWebEngineView and hide
tooltips only if requested by Chromium.

Fixes: QTBUG-97897
Pick-to: 6.2
Change-Id: I976e49218ec975b10e2f155096f6aec124270fdc
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>