Doc: Remove broken homepage link from markdowncss attribution
QtWebEngine uses Keving Burke's markdowncss for two of his examples.

An attribution file is provided for both uses, pointing to the homepage
of markdowncss "https://kevinburke.bitbucket.io/markdowncss/".

Nonetheless, the project is currently unmaintained and the homepage
doesn't exist anymore, meaning that the link to it is broken.

As there is not a reachable homepage anymore, the "Homepage" key for
markdowncss attributions was removed from the attribution files.

A link to the same page was changed to the link to the bitbucket
repository, which is still reachable, in the documentation for the
affected examples.

Task-number: QTBUG-96127
Change-Id: Idd6c163eebeb2a2292ae76a73289fda5faa12b0a
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
(cherry picked from commit 222b3b054cfda31b3bbf3059e3e53139d00bb76f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>