Fix crash with disabled geolocation on macOS
QtWebEngine uses custom LocationProvider for geolocation. If geolocation
is disabled, Chromium fallbacks to NetworkLocationProvider.

NetworkLocationProvider uses GeolocationSystemPermissionManager and
tThere is no nullptr check for the manager instance so
QtWebEngine has to provide it. This fix implements a
FakeSystemGeolocationPermissionManager what is meant to be used only
on macOS if geolocation is disabled.

LocationSystemPermissionStatus is always set to denied by the fake manager.

Pick-to: 6.2
Change-Id: I25d51c9ce8911b95ff69cc72bc6aae7023e7edbe
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>