Fix not working web ui popup (html comboboxes) due to focus out events
Recent change in qtdeclarative 42d411e2e8 causes a focus out event
being delivered in case of popups to "root" item of quickwidget.

This is not expected as events are forwarded to parent "view" and
loosing a focus will hint Blink to trigger a pop up close request.

As a communication with a render process is asynchronous this
creates several race conditions in our tests and the Blink's popup
close request can close unexpectedly windows creating dangling pointers.

Moreover, the focus in never gained back leaving the Blink's logic in
a limbo state.

Simply ignore a focus out event in case of popups.

Fixes: QTBUG-99215
Pick-to: 6.3 6.2 6.2.3
Change-Id: I5ca6eda227101d4f19f15735e41f066cfd8ccea0
Reviewed-by: Kirill Burtsev <kirill.burtsev@qt.io>