Introduce gn complete static lib target for qtpdf
With 6.x we have gn-cmake integration which allows now
in case of static builds to switch gn target output to
static_library with 'complete static' option instead of
shared_library. This way we can avoid installing dependencies
in form of internal static Chromium archives as it was done
in 5.15 in case of static builds.

Unfortunately 'complete static' option can not be used with
QtWebEngineCore build as it will exceed 4Gb static archive
limitation on Windows, however it will work fine with QtPdf.

Note the qtbase 3rdparty static libs have to be still installed
in case of qt static build, this patch only eliminates need for
Chromium specific ones.

Task-number: QTBUG-88614
Pick-to: 6.3 6.2
Change-Id: I6fc4ce48c79a1631b013d0b29c190c62280f7304
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>