QQuickPdfSelection: Fix division by zero
QQuickPdfSelection::m_renderScale is used for division in
QQuickPdfSelection::updateResults(). Add warning message to
setRenderScale() and ignore the new value if it is zero.

Also avoid setting PdfSelection.renderScale to zero in
PdfScrollablePageView.

The division caused an assert:
ASSERT: "divisor < 0 || divisor > 0" in file qtbase/src/corelib/tools/qpoint.h, line 299

It was reproducible with the pdfviewer quick example.

Pick-to: 6.3 6.2
Change-Id: I31e08a9a28392fcc132470c8d94af1307447565a
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>