doc: Provide a screenshot of the PDF Viewer widget example; rename page
Let's be clear in the title that this is a widget-based example, as
opposed to the Qt Quick example in the docs.

The screenshot is from macOS 13.2.1.

The 6.6 example has more features, so this screenshot looks a bit
incomplete compared to that.

Task-number: QTBUG-81560
Change-Id: I28b50d608756f10967f1bf5f7ac72ec9a83d9417
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>
(cherry picked from commit 9937cbb4a0f43a799400f4057e8c695f88d1aa9f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>