CMake: Adjust qt_add_webengine_dictionary API
Rename the webengine_dictionaries custom target to
qtwebengine_dictionaries.

We don't name it qt_webengine_dictionaries for it to be consistent
with the name of the directory where WebEngine looks up the binary
dictionaries.

Also fix the docs to mention the correct location of the dictionary
output directory.

Pick-to: 6.3
Task-number: QTBUG-100210
Change-Id: I0232fa17d904ff602a09a70d21ad3d94c145b4c7
Reviewed-by: Michal Klocek <michal.klocek@qt.io>