QtQuick.Pdf: automatically close the error dialog if the error disappears
When the user has not yet entered a password, document.status == Error;
but we pop up the password prompt dialog on top of the error dialog.
After the user has entered it, either it's correct and the error goes
away, or we need to tell the user that it's incorrect, so the dialog
should stay open.

Change-Id: Ib3332f543b23fdf50ccd0df0c7e79cdc95f2e8da
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>