QtPdf: hide "debug" properties; qt.pdf.(single|multi]pageview log cats
The property "debug" was not nice API, but I have not gotten in the
habit of using logging categories from QML. This is much nicer.

The "debug" property is still useful to turn on some visual debugging
features, but it's better to hide it rather than keeping it exposed and
needing to document it.

Pick-to: 6.3
Change-Id: I5ec4156e6d4e7ccd0f406fe62ee810f4401561ec
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>